Feature/change float to double

+++ Color.cpp
@@ -7,12 +7,12 @@
 #include "include/Color.h"
 
 Color::Color() {
-    R = 0.f;
-    G = 0.f;
-    B = 0.f;
+    R = 0.;
+    G = 0.;
+    B = 0.;
 }
 
-Color::Color(float R, float G, float B)
+Color::Color(double R, double G, double B)
 {
     this->R = R;
     this->G = G;
@@ -40,37 +40,37 @@ std::ostream &operator<<(std::ostream &os, const Color &C)
 
 TEST_CASE("adding colors")
 {
-    Color A = Color(0.9f, 0.6f, 0.75f);
-    Color B = Color(0.7f, 0.1f, 0.25f);
-    CHECK((A + B) == Color(1.6f, 0.7f, 1.0f));
+    Color A = Color(0.9, 0.6, 0.75f);
+    Color B = Color(0.7, 0.1, 0.25f);
+    CHECK((A + B) == Color(1.6, 0.7, 1.0f));
 }
 
 TEST_CASE("subtracting colors")
 {
-    Color A = Color(0.9f, 0.6f, 0.75f);
-    Color B = Color(0.7f, 0.1f, 0.25f);
-    CHECK((A - B) == Color(0.2f, 0.5f, 0.5f));
+    Color A = Color(0.9, 0.6, 0.75f);
+    Color B = Color(0.7, 0.1, 0.25f);
+    CHECK((A - B) == Color(0.2, 0.5, 0.5f));
 }
 
 TEST_CASE("multiplying a color by a scalar")
 {
-    Color A = Color(0.2f, 0.3f, 0.4f);
-    CHECK((A * 2) == Color(0.4f, 0.6f, 0.8f));
-    CHECK((2 * A) == Color(0.4f, 0.6f, 0.8f));
+    Color A = Color(0.2, 0.3, 0.4f);
+    CHECK((A * 2) == Color(0.4, 0.6, 0.8f));
+    CHECK((2 * A) == Color(0.4, 0.6, 0.8f));
 }
 
 TEST_CASE("multiplying color")
 {
-    Color A = Color(1.f, 0.2f, 0.4f);
-    Color B = Color(0.9f, 1.f, 0.1f);
-    CHECK((A * B) == Color(0.9f, 0.2f, 0.04f));
+    Color A = Color(1., 0.2, 0.4f);
+    Color B = Color(0.9, 1., 0.1f);
+    CHECK((A * B) == Color(0.9, 0.2, 0.04f));
 }
 
 TEST_CASE("convert color to ppm string")
 {
-    Color A = Color(1.5f, 0.f, 0.f);
-    Color B = Color(0.f, 0.5f, 0.f);
-    Color C = Color(-0.5f, 0.f, 1.f);
+    Color A = Color(1.5, 0., 0.f);
+    Color B = Color(0., 0.5, 0.f);
+    Color C = Color(-0.5, 0., 1.f);
     std::vector<int> APPM = {255, 0, 0};
     std::vector<int> BPPM = {0, 128, 0};
     std::vector<int> CPPM = {0, 0, 255};
