+++ Ray.cpp
@@ -36,7 +36,7 @@ Vector Ray::GetDirection() const
     return Direction;
 }
 
-Point Ray::Position(float T)
+Point Ray::Position(double T)
 {
     return Origin + Direction * T;
 }
@@ -53,38 +53,38 @@ Ray Ray::Transform(Matrix &&M) const
 
 TEST_CASE("Creating and querying a Ray")
 {
-    Ray R(Point(1.f, 2.f, 3.f), Vector(4.f, 5.f, 6.f));
+    Ray R(Point(1., 2., 3.f), Vector(4., 5., 6.f));
 
-    CHECK(R.GetOrigin() == Point(1.f, 2.f, 3.f));
-    CHECK(R.GetDirection() == Vector(4.f, 5.f, 6.f));
+    CHECK(R.GetOrigin() == Point(1., 2., 3.f));
+    CHECK(R.GetDirection() == Vector(4., 5., 6.f));
 }
 
 TEST_CASE("Computing a point from a distance")
 {
-    Ray R(Point(2.f, 3.f, 4.f), Vector(1.f, 0.f, 0.f));
+    Ray R(Point(2., 3., 4.f), Vector(1., 0., 0.f));
 
-    CHECK(R.Position(0.f) == Point(2.f, 3.f, 4.f));
-    CHECK(R.Position(1.f) == Point(3.f, 3.f, 4.f));
-    CHECK(R.Position(-1.f) == Point(1.f, 3.f, 4.f));
-    CHECK(R.Position(2.5f) == Point(4.5f, 3.f, 4.f));
+    CHECK(R.Position(0.f) == Point(2., 3., 4.f));
+    CHECK(R.Position(1.f) == Point(3., 3., 4.f));
+    CHECK(R.Position(-1.f) == Point(1., 3., 4.f));
+    CHECK(R.Position(2.5f) == Point(4.5, 3., 4.f));
 }
 
 TEST_CASE("Translating a ray")
 {
-    Ray R(Point(1.f, 2.f, 3.f), Vector(0.f, 1.f, 0.f));
-    auto M = Matrix::Translation(3.f, 4.f, 5.f);
+    Ray R(Point(1., 2., 3.f), Vector(0., 1., 0.f));
+    auto M = Matrix::Translation(3., 4., 5.f);
     auto R2 = R.Transform(M);
 
-    CHECK(R2.GetOrigin() == Point(4.f, 6.f, 8.f));
-    CHECK(R2.GetDirection() == Vector(0.f, 1.f, 0.f));
+    CHECK(R2.GetOrigin() == Point(4., 6., 8.f));
+    CHECK(R2.GetDirection() == Vector(0., 1., 0.f));
 }
 
 TEST_CASE("Scaling a ray")
 {
-    Ray R(Point(1.f, 2.f, 3.f), Vector(0.f, 1.f, 0.f));
-    auto M = Matrix::Scaling(2.f, 3.f, 4.f);
+    Ray R(Point(1., 2., 3.f), Vector(0., 1., 0.f));
+    auto M = Matrix::Scaling(2., 3., 4.f);
     auto R2 = R.Transform(M);
 
-    CHECK(R2.GetOrigin() == Point(2.f, 6.f, 12.f));
-    CHECK(R2.GetDirection() == Vector(0.f, 3.f, 0.f));
+    CHECK(R2.GetOrigin() == Point(2., 6., 12.f));
+    CHECK(R2.GetDirection() == Vector(0., 3., 0.f));
 }
